Add board enable for ASUS P5LP-LE
Based on the REing of Michael Karcher and a good guess of Idwer Vollering.

Signed-off-by: Michael Karcher <flashrom@mkarcher.dialup.fu-berlin.de>
Signed-off-by: Stefan Tauner <stefan.tauner@student.tuwien.ac.at>
Tested-by: Leon Zhang <leoncamel at gmail.com>
http://www.coreboot.org/pipermail/flashrom/2010-August/004500.html
Tested-by: medhi <nefkongo@hotmail.com>
http://paste.flashrom.org/view.php?id=779
Acked-by: Idwer Vollering <vidwer@gmail.com>
---
i have added the DMI match and changed the board enable and wiki comments
a bit. i am not sure if this is the right way to add it, because there are so many
variations, but we can change that later too.