layout: Verify layout entries before building a new image using them.
This fixes a SEGFAULT if a layout entry is included that addresses memory
outside the current chip's address range. flashrom will only abort if the
offending region(s) is/are included else it will just warn.

It will print warnings for regions with negative or zero-length address
ranges too, but it will only abort if they are included with -i/--image to reduce
the potential of regressions. This is different to the patch committed to
the development branch in r1751.

Also, abort for non-write operations if a layout file is given because
there is no layout support for non-write operations yet, and some reports
show that users expect it to work at least for -r/--read.

Signed-off-by: Stefan Tauner <stefan.tauner@student.tuwien.ac.at>
Acked-by: Carl-Daniel Hailfinger <c-d.hailfinger.devel.2006@gmx.net>